Extract getNetwork in bitcoin
