+++ packages/snap/snap.manifest.json
@@ -7,7 +7,7 @@
     "url": "https://github.com/KeystoneHQ/btcsnap"
   },
   "source": {
-    "shasum": "O02MMukttFtc3G4u4QCfTXXOn7O3uDZlSMJSWbMmOJ4=",
+    "shasum": "JTn2I5qRLU9XFVhZ1Ep88jCDREpa2Vxv58Kt0SwM68M=",
     "location": {
       "npm": {
         "filePath": "dist/bundle.js",
